[TAT.hpp] do not remove unused symmetry by default
it is only called when edge operator, because when
merging, it ususally generate many unused symmetry.
that is the only reason we need to remove unused
symmetry.

In the other hand, in svd, qr and contract, we call
edge operator first to generate matrix and use their
edge to do loop operator, so their edge should not
have extra useless information, call remove unused
symemtry in edge operator ensure every temporary
matrix in svd, qr and cotnract does not have unused
symmetry.