[TAT.hpp] update interface of tensor conjugate.
The behavior of physics edge and virtual edge when conjugating is
different. The original positive_contract is a only condition if all
edges are physics edge.

Now, c++ api can treat more detailed condition, if some edges are
physics and somes are not.

Usually, arrow of physics edges is False, while for arrow=False edge,
the behavior between physics edge and virtual edge are the same. so
treat those physics edge as virtual edge is ok.