[tet.py] rename parameter of gradient descent `sweep_initial_configuration` to `sampling_configurations`.
since it is also used for direct sampling.

its behavior is, whatever sampling used, the last sampling
configuration will be stored in it. While for sweep sampling, it will
use this as initial configuration. If this argument is missing, an
empty list will be used, and of course for sweep sampling it will crash.