[example] add density matrix of heisenberg as individual model file.
Density matrix of heisenberg is just a normal model now, and it is
easy to do simple update or natural gradient on it. An example is
added into /examples/heisenberg_gibbs_state/ which calls the
initializer for density matrix especially, since density matrix should
be initialized as identity for infinite temperature state.