[TAT.hpp] fix wrong constructor when pass iterator to symmetry list.
for nosymmetry edge, which is constructed by an int as dimension, if I
want to pass a vector of int to construct tensor, it will be converted
to vector of edge first. for example, program may try to convert
`{a.begin(), a.end()}` to vector of edge from vector of int. however,
the original version of code accept any type to try to construct edge,
so it will try to convert `a.begin()` to a single edge, which is not
what expected.