[TAT.hpp] Fix a bug when copying fermi edge.
This commit is related to commit
637674d6706f34b20d2238948bbe9bdf65e350c8.

Originally, Using `not self` to check whether to use the constructor
of edge. But if an iterator given, which is expected to construct a
series of edges, could be considered to construct a single edge by
compiler. So at the commit 637674d6706f34b20d2238948bbe9bdf65e350c8,
the condition was changed to check if it is constructable for base
class of edge.

However, the derived class is also constructable for base class of
edge. And on windows platform, it will use this constructor even if
args is Edge itself, whichh is not expected.

So now, using the two condition to double check whether to use the
constructor.