[TAT.hpp] Print blocks of tensor in new order.
The old order is the dictionary order of symmetry list for the block.
The new order is the dictionary order of position list for the block.

Only the print order is updated. The order of storage is still in the old order
for compatibility. This will be updated to new order later.