[TAT.hpp] Add svd cut method: boltzmann cut. (#29)
Which use a "temperature" to cut dimension, if temperature=0, it equals to
remain cut. In higher temperature, it has more possibility to choose a smaller
singular value, This may be useful for blocked tensor, which erasing a symmetry
may be much more harmful than cutting a larger singular. Inside a single block,
smaller singular is always cutted earlier than larger one, so this commit does
not affect no symmetry tensor.