[tet.py] Fix min-SR bug when calculating fermionic system.
The delta calculated by scalapack is in ket space.
But when converting delta to array, it is in bra space.
So conjugate the array only to move it to bra space again,
only then it is valid to use it as tensor storage. After
converting it to tensors, conjugate the tensor to get the
natural gradient.