[tetku.py] Revert "Gibbs state second layer is total reversed, not half."
This reverts commit f006f748a2a88575d52bf4abb10928e8ece586b6.

As shown below, there is totally 5 edge need to be reversed in fact.
each edge produce a sign, so a single sign is produced totally, that
is say, only apply half of edges in Hamiltonian when reverse.

                virtual bond
[tensor at H] ------------------ [other part of psi]
      |                                   |
[     H     ]                             | this edge contain id and T
      |         virtual bond              |
[tensor at H]------------------- [other part of psi]

When merge or split edge on Hamiltonian, it is also needed to applying
sign to half of edges in Hamiltonian, but their reason is totally
different: because of the noncommutability between merge/split and
conjugate.