TAT v0.3.9
Some messages and branches in git history before TAT v0.3.9 are in a
mess, so let's rebuild a new history. The previous history could also
be found in the old branch, named v0.3.9.

When migrating to this new git history tree, the only change is the
update of code format style. .clang-format is updated, previous indent
width=3 is wired, and now it is 4.