[build] Push to main branch automatically when a tag pushed to github.
The main branch is designed to be stable, it will should be updated
at the last step of github action `build` automatically.

The future workflow is like: feature or bugfix squashed into dev, when
a release is needed, tag the commit on dev and push the tag to github,
after all the check passed, gitbot push the tag to main branch.