[tet.py] Update the configuration persistent data structure.
To achieve this update, the physics edge index should be non-negative
now. The previous persistent configuration is a nested python
list/dict object, which is slow to allgather. The new structure make a
3-d numpy array with `dtype=np.int64`, which shape is L1 * L2 *
max_orbit. Inside the array, the edge_index is stored, if it is None,
it would be -2, if the edge of the current index does not exist, it
would be -1.

gm_conf become a numpy array instead of python list. Previously, it
relies on passing by ref. so some corresponding update has been
applied. Add copy to gm_conf_load to let owndata flag be true.