[TAT.hpp] Remove type for template alias of blas/lapack function.
In fact, the type for template alias does not affect the
specialization, while in TAT's situation, it always use the
specialized alias, so, default type of template alias is useless, and
now it is removed. Besides, add `constexpr` specifier to the
specialization, since the template default specifier seems not affect
the specialized value.