[tetaux] Remove tetraux.
It is useless since ansatz product state removed. Previously, the only
function in tetraux is the configuration object used in ansatz product
state. In fact, the new configuration structure for sampling lattice
has the similar design. If we bring ansatz product state back in the
future, we can use something like configuration of sampling lattice,
which is a numpy array, instead of implementing it in c++ vector with
pybind11.