[tetku.py] Fix bug in particle hopping hamiltonian for gibbs hubbard.
For infinite temperature state, all non zero configuration requires it
is diagonal. But the previous hopping need to pass by a non diagonal
configuration to reach another diagonal configurtation.

This commit add another hopping term which can make particle in each
layer hopping to nearest site at the same time. So it keep the
connectedness even in the infinite temperature.