[build] Add support for musllinux platforms in github action.
The cibuildwheel supports many architecture and system.

For amd64, which is just github action architecture, the supported
system are macos, win, manylinux and musllinux, this commit make all
of them supported.

For arm64, our only supported system is macos, which is easy to build.
While for linux, the default compiler is gcc but not clang, cross
compiling is hard to configure. And for windows, although we can use
clang to cross compile our package, but openblas does not provide
prebuilt library.

For other architecture, we never use them, so do not build them
currently.