[build] Add support for linux arm64 in github action building PyTAT.
The package cibuildwheel gives a solution to use qemu docker to
simulate arm64 architecture. And it is designed to create python wheel
in architectures other than amd64, so use it.

After this commit, the prebuilt wheel for PyTAT is shown in the table
below:
|-------+-------+---------+-----------+-----------|
| os    | macos | windows | manylinux | musllinux |
|-------+-------+---------+-----------+-----------|
| amd64 | OK    | OK      | OK        | OK        |
| arm64 | OK    |         | OK        | OK        |
|-------+-------+---------+-----------+-----------|
Only windows arm64 is not supported.