[build, TAT.py] Fix dependency of pytat in emscripten wheel.
This commit try use `pyodide auditwheel` to embed openblas into pytat
emscripten wheel.
But this remain in comment since upstream does not support it yet.
When everything ready just uncomment it.
And inside `test_wasm.js`, we use micropip.install instead of
pyodide.loadPackage to handle the dependency correctly.
The only thing it cannot handle is the openblas, we load it manually
temporarily.

(close #40)